refactor(utils): change util files location to src/util and minor readability changes
